Create output file for executable to write to
The executable cannot create any new files to write to, so it needs a blank default one that it can write information to. It will overwrite the file with new data for every execution. The data will then be recorded to a newly created and formatted file via the C# code since the python does not have that capability as an executable.